Extract inner classes from the CentralJobScheduler.
