+++ community/kernel/src/main/java/org/neo4j/kernel/impl/scheduler/RegisteredPooledJobHandle.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2002-2018 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.kernel.impl.scheduler;
+
+import java.util.Set;
+import java.util.concurrent.Future;
+
+import org.neo4j.scheduler.JobScheduler;
+
+final class RegisteredPooledJobHandle extends PooledJobHandle
+{
+    private final Set<JobScheduler.JobHandle> register;
+
+    RegisteredPooledJobHandle( Future<?> job, Set<JobScheduler.JobHandle> register )
+    {
+        super( job );
+        this.register = register;
+    }
+
+    @Override
+    public void cancel( boolean mayInterruptIfRunning )
+    {
+        super.cancel( mayInterruptIfRunning );
+        register.remove( this );
+    }
+}
